src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albumentations==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
augraphy==8.2.6
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.5
bleach==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dotenv==0.9.9
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
fonttools==4.59.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
gitpython==3.1.45
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
ipykernel==6.27.1
ipython==8.15.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter==1.0.0
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.9
lark==1.2.2
lazy-loader==0.4
lightning==2.5.5
lightning-utilities==0.15.2
llvmlite==0.44.0
markupsafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.6
mistune==3.1.4
mpmath==1.3.0
multidict==6.6.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.5
notebook-shim==0.2.4
numba==0.61.2
numpy==1.26.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.5
pathlib==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==9.4.0
pip==25.0.1
platformdirs==4.4.0
prometheus-client==0.22.1
prompt-toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytorch-lightning==2.5.5
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.2
qtconsole==5.6.1
qtpy==2.4.3
qudida==0.0.4
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-image==0.24.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
send2trash==1.8.3
sentry-sdk==2.36.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.8.28
timm==0.9.12
tinycss2==1.4.0
torch==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
types-python-dateutil==2.9.0.20250822
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uv==0.8.14
wandb==0.21.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
yarl==1.20.1
